split out distribution
